refactor: reorganizes the structure of the packages
